+++ tybotRoute/tiledeskChatbotPlugs/DirectivesChatbotPlug.js
@@ -459,7 +459,7 @@ class DirectivesChatbotPlug {
       });
     }
     else if (directive_name === Directives.WEB_REQUEST) {
-      console.log("...DirWebRequest");
+      // console.log("...DirWebRequest");
       new DirWebRequest(context).execute(directive, async () => {
         let next_dir = await this.nextDirective(this.directives);
         this.process(next_dir);